+++ src/components/Signup/Emailsignup/emailsignup.tsx
@@ -1,5 +1,4 @@
-import React, { useState } from 'react';
-import axios from 'axios';
+import React from 'react'
 import * as S from '@/components/Signup/Emailsignup/emailsignup.style';
 import hidePasswordimg from '@/assets/image/onbording/hide_fill.svg';
 import showPasswordimg from '@/assets/image/onbording/show_fill.svg';
@@ -7,112 +6,12 @@ import Cloud1 from '@/assets/image/onbording/oauthsignup/Cloud1.svg';
 import Cloud2 from '@/assets/image/onbording/oauthsignup/Cloud2.svg';
 import Cloud3 from '@/assets/image/onbording/oauthsignup/Cloud3.svg';
 import Sun from '@/assets/image/onbording/oauthsignup/Sun.svg'
-import config from '@/config/config.json';
 import Button from '@/components/button/Button';
-import { useNavigate } from 'react-router';
 import TextField from '@/components/TextField/TextField';
+import useSignup from '@/hooks/Signuphook/Signup';
 
 const emailsignup = () => {
-    const navigate = useNavigate();
-    const [showPassword, setShowPassword] = useState<boolean>(false);
-    const [showConfirmPassword, setShowConfirmPassword] = useState<boolean>(false);
-    const [password, setPassword] = useState<string>('');
-    const [confirmPassword, setConfirmPassword] = useState<string>('');
-    const [errorMessage, setErrorMessage] = useState<string>('');
-    const [name, setName] = useState<string>('');
-    const [email, setEmail] = useState<string>('');
-
-    const handleNameChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-        setName(e.target.value);
-    };
-
-    const handleEmailChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-        setEmail(e.target.value);
-    }
-
-    const togglePasswordVisibility = () => {
-        setShowPassword(!showPassword);
-    };
-
-    const toggleConfirmPasswordVisibility = () => {
-        setShowConfirmPassword(!showConfirmPassword);
-    };
-
-    const handlePasswordChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-        setPassword(e.target.value);
-    };
-
-    const handleConfirmPasswordChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-        setConfirmPassword(e.target.value);
-    };
-
-    const validateEmail = (email: string) => {
-        const re = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-        return re.test(String(email).toLowerCase());
-    }
-
-    const validatePassword = (password: string) => {
-        const re = /^(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,}$/;
-        return re.test(password);
-    };
-
-    
-    const clearErrorMessage = () => {
-        setErrorMessage('');
-    }
-    const handleSignup = async () => {
-        if (!name.trim()) {
-            setErrorMessage('이름을 입력해주세요');
-            setTimeout(clearErrorMessage, 3000);
-            return;
-        }
-        if (!email.trim()) {
-            setErrorMessage('이메일을 입력해주세요.');
-            setTimeout(clearErrorMessage, 3000);
-            return;
-        }
-        if (!validateEmail(email)) {
-            setErrorMessage('유효한 이메일 형식을 입력해주세요.');
-            setTimeout(clearErrorMessage, 3000);
-            return;
-        }
-        if (!password.trim()) {
-            setErrorMessage('비밀번호를 입력해주세요.');
-            setTimeout(clearErrorMessage, 3000);
-            return;
-        }
-        if (password !== confirmPassword) {
-            setErrorMessage('비밀번호가 일치하지 않습니다.');
-            setTimeout(clearErrorMessage, 3000);
-        } 
-        if(!validatePassword(password)) {
-            setErrorMessage('비밀번호는 8자리 이상, 특수문자 포함이어야 합니다.'); 
-            setTimeout(clearErrorMessage, 3000);
-        } else {
-            try {
-                const response = await axios.post(`${config.serverurl}/member/register`, {
-                    name,
-                    email,
-                    password,
-                    // 토큰 추가
-                });
-                if (response.status === 200) {
-                    navigate('/emailathentance');
-                } else {
-                    alert('회원가입 중 문제가 발생했습니다. 다시 시도해주세요.');
-                }
-            } catch (error) {
-                console.error('error:', error);
-                alert('회원가입 중 문제가 발생했습니다. 다시 시도해주세요.');
-            }
-        }
-    };
-
-    const handleKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
-        if (e.key === 'Enter') {
-            handleSignup();
-        }
-    }
+    const { ...signUp } = useSignup();
 
     return (
         <S.EmailMain>
@@ -132,9 +31,9 @@ const emailsignup = () => {
                         <S.InputContainer>
                             <TextField
                                 placeholder='이름을 입력해주세요'
-                                onChange={handleNameChange}
+                                onChange={signUp.handleNameChange}
                                 style={{ border: "none" }}
-                                onKeyDown={handleKeyDown}
+                                onKeyDown={signUp.handleKeyDown}
                                 />
                         </S.InputContainer>
                     </S.EneterInfo>
@@ -146,8 +45,8 @@ const emailsignup = () => {
                             <TextField
                                 style={{ border: "none" }}
                                 placeholder='이메일을 입력해주세요'
-                                onChange={handleEmailChange}
-                                onKeyDown={handleKeyDown} 
+                                onChange={signUp.handleEmailChange}
+                                onKeyDown={signUp.handleKeyDown} 
                                 />
                         </S.InputContainer>
                     </S.EneterInfo>
@@ -158,13 +57,13 @@ const emailsignup = () => {
                         <S.InputContainer>
                             <TextField
                                 style={{ border: "none" }}
-                                type={showPassword ? 'text' : 'password'}
+                                type={signUp.showPassword ? 'text' : 'password'}
                                 placeholder='비밀번호를 입력해주세요'
-                                onChange={handlePasswordChange}
-                                onKeyDown={handleKeyDown}
+                                onChange={signUp.handlePasswordChange}
+                                onKeyDown={signUp.handleKeyDown}
                             />
-                            <S.Btnview onClick={togglePasswordVisibility}>
-                                {showPassword ? <img src={hidePasswordimg} alt="숨기기" /> : <img src={showPasswordimg} alt="보이기" />}
+                            <S.Btnview onClick={signUp.togglePasswordVisibility}>
+                                {signUp.showPassword ? <img src={hidePasswordimg} alt="숨기기" /> : <img src={showPasswordimg} alt="보이기" />}
                             </S.Btnview>
                         </S.InputContainer>
                     </S.EneterInfo>
@@ -175,20 +74,20 @@ const emailsignup = () => {
                         <S.InputContainer>
                             <TextField
                                 style={{ border: "none" }}
-                                type={showConfirmPassword ? 'text' : 'password'}
+                                type={signUp.showConfirmPassword ? 'text' : 'password'}
                                 placeholder='비밀번호를 다시 입력해주세요'
-                                onChange={handleConfirmPasswordChange}
-                                onKeyDown={handleKeyDown}
+                                onChange={signUp.handleConfirmPasswordChange}
+                                onKeyDown={signUp.handleKeyDown}
                             />
-                            <S.Btnview onClick={toggleConfirmPasswordVisibility}>
-                                {showConfirmPassword ? <img src={hidePasswordimg} alt="숨기기" /> : <img src={showPasswordimg} alt="보이기" />}
+                            <S.Btnview onClick={signUp.toggleConfirmPasswordVisibility}>
+                                {signUp.showConfirmPassword ? <img src={hidePasswordimg} alt="숨기기" /> : <img src={showPasswordimg} alt="보이기" />}
                             </S.Btnview>
                         </S.InputContainer>
-                        {errorMessage && <S.ErrorText>{errorMessage}</S.ErrorText>}
+                        {signUp.errorMessage && <S.ErrorText>{signUp.errorMessage}</S.ErrorText>}
                     </S.EneterInfo>
                 </S.TxtContainer>
                 <S.ButtonContainer>
-                    <Button onClick={handleSignup}/>
+                    <Button onClick={signUp.handleSignup}/>
                     <S.EmailCheck>
                         <S.Haveemail href='http://localhost:5173/login'>이미 계정이 있으신가요?</S.Haveemail>
                     </S.EmailCheck>
